refactor(xray): move structs and functions to utils file
