+++ xray/services/xraywatch.go
@@ -4,24 +4,26 @@ import (
 	"encoding/json"
 	"errors"
 	"net/http"
-	"sort"
 
 	rthttpclient "github.com/jfrog/jfrog-client-go/artifactory/httpclient"
-	"github.com/jfrog/jfrog-client-go/artifactory/services/utils"
+	// "github.com/jfrog/jfrog-client-go/artifactory/services/utils"
+
+	artUtils "github.com/jfrog/jfrog-client-go/artifactory/services/utils"
 	"github.com/jfrog/jfrog-client-go/auth"
 	clientutils "github.com/jfrog/jfrog-client-go/utils"
 	"github.com/jfrog/jfrog-client-go/utils/errorutils"
 	"github.com/jfrog/jfrog-client-go/utils/log"
+	"github.com/jfrog/jfrog-client-go/xray/services/utils"
 )
 
 const (
-	WatchBuildAll    WatchBuildType = "all"
-	WatchBuildByName WatchBuildType = "byname"
-
-	WatchRepositoriesAll    WatchRepositoriesType = "all"
-	WatchRepositoriesByName WatchRepositoriesType = "byname"
-
 	WatchAPIURL = "api/v2/watches"
+
+	WatchBuildAll    utils.WatchBuildType = "all"
+	WatchBuildByName utils.WatchBuildType = "byname"
+
+	WatchRepositoriesAll    utils.WatchRepositoriesType = "all"
+	WatchRepositoriesByName utils.WatchRepositoriesType = "byname"
 )
 
 type XrayWatchService struct {
@@ -57,8 +59,8 @@ func (xws *XrayWatchService) Delete(watchName string) error {
 	return nil
 }
 
-func (xws *XrayWatchService) Create(params XrayWatchParams) error {
-	payloadBody, err := CreateBody(params)
+func (xws *XrayWatchService) Create(params utils.XrayWatchParams) error {
+	payloadBody, err := utils.CreateBody(params)
 	if err != nil {
 		return errorutils.CheckError(err)
 	}
@@ -69,7 +71,7 @@ func (xws *XrayWatchService) Create(params XrayWatchParams) error {
 	}
 
 	httpClientsDetails := xws.XrayDetails.CreateHttpClientDetails()
-	utils.SetContentType("application/json", &httpClientsDetails.Headers)
+	artUtils.SetContentType("application/json", &httpClientsDetails.Headers)
 	var url = xws.GetXrayWatchUrl()
 	var resp *http.Response
 	var respBody []byte
@@ -92,170 +94,8 @@ func (xws *XrayWatchService) Create(params XrayWatchParams) error {
 	return nil
 }
 
-func CreateBody(params XrayWatchParams) (*XrayWatchBody, error) {
-	payloadBody := XrayWatchBody{
-		GeneralData: XrayWatchGeneralParams{
-			Name:        params.Name,
-			Description: params.Description,
-			Active:      params.Active,
-		},
-		ProjectResources: XrayWatchProjectResources{
-			Resources: []XrayWatchProjectResourcesElement{},
-		},
-		AssignedPolicies: params.Policies,
-	}
-
-	err := ConfigureRepositories(&payloadBody, params)
-	if err != nil {
-		return nil, err
-	}
-
-	err = ConfigureBuilds(&payloadBody, params)
-	if err != nil {
-		return nil, err
-	}
-
-	err = ConfigureBundles(&payloadBody, params)
-	if err != nil {
-		return nil, err
-	}
-
-	return &payloadBody, nil
-}
-
-func ConfigureRepositories(payloadBody *XrayWatchBody, params XrayWatchParams) error {
-	if params.Repositories.Type == WatchRepositoriesAll {
-		allFilters := XrayWatchProjectResourcesElement{
-			Type:          "all-repos",
-			StringFilters: []XrayWatchFilter{},
-		}
-
-		allFilters.StringFilters = append(allFilters.StringFilters, CreateFilters(params.Repositories.All.Filters, params.Repositories)...)
-
-		payloadBody.ProjectResources.Resources = append(payloadBody.ProjectResources.Resources, allFilters)
-	} else if params.Repositories.Type == WatchRepositoriesByName {
-		for _, repository := range params.Repositories.Repositories {
-			repo := XrayWatchProjectResourcesElement{
-				Type:          "repository",
-				Name:          repository.Name,
-				BinMgrID:      repository.BinMgrID,
-				StringFilters: repository.StringFilters,
-			}
-			if repo.StringFilters == nil {
-				repo.StringFilters = []XrayWatchFilter{}
-			}
-			repo.StringFilters = append(repo.StringFilters, CreateFilters(repository.Filters, params.Repositories)...)
-
-			payloadBody.ProjectResources.Resources = append(payloadBody.ProjectResources.Resources, repo)
-		}
-	}
-
-	return nil
-}
-
-func CreateFilters(filters XrayWatchFilters, repo XrayWatchRepositoriesParams) []XrayWatchFilter {
-	result := []XrayWatchFilter{}
-
-	for _, packageType := range filters.PackageTypes {
-		filter := XrayWatchFilter{
-			Type:  "package-type",
-			Value: packageType,
-		}
-		result = append(result, filter)
-	}
-
-	for _, name := range filters.Names {
-		filter := XrayWatchFilter{
-			Type:  "regex",
-			Value: name,
-		}
-		result = append(result, filter)
-	}
-
-	for _, path := range filters.Paths {
-		filter := XrayWatchFilter{
-			Type:  "path-regex",
-			Value: path,
-		}
-		result = append(result, filter)
-	}
-
-	for _, mimeType := range filters.MimeTypes {
-		filter := XrayWatchFilter{
-			Type:  "mime-type",
-			Value: mimeType,
-		}
-		result = append(result, filter)
-	}
-
-	for key, value := range filters.Properties {
-		filter := XrayWatchFilter{
-			Type: "property",
-			Value: XrayWatchFilterPropertyValue{
-				Key:   key,
-				Value: value,
-			},
-		}
-		result = append(result, filter)
-	}
-
-	if repo.ExcludePatterns != nil || repo.IncludePatterns != nil {
-		filter := XrayWatchFilter{
-			Type: "path-ant-patterns",
-			Value: XrayWatchPathFilters{
-				ExcludePatterns: repo.ExcludePatterns,
-				IncludePatterns: repo.IncludePatterns,
-			},
-		}
-		result = append(result, filter)
-	}
-
-	return result
-}
-
-func ConfigureBuilds(payloadBody *XrayWatchBody, params XrayWatchParams) error {
-	if params.Builds.Type == WatchBuildAll {
-		allBuilds := XrayWatchProjectResourcesElement{
-			Name:          "All Builds",
-			Type:          "all-builds",
-			BinMgrID:      params.Builds.All.BinMgrID,
-			StringFilters: []XrayWatchFilter{},
-		}
-
-		if params.Builds.All.ExcludePatterns != nil || params.Builds.All.IncludePatterns != nil {
-			filters := []XrayWatchFilter{{
-				Type: "ant-patterns",
-				Value: XrayWatchPathFilters{
-					ExcludePatterns: params.Builds.All.ExcludePatterns,
-					IncludePatterns: params.Builds.All.IncludePatterns,
-				}},
-			}
-			allBuilds.StringFilters = filters
-		}
-
-		payloadBody.ProjectResources.Resources = append(payloadBody.ProjectResources.Resources, allBuilds)
-	} else if params.Builds.Type == WatchBuildByName {
-		for _, byName := range params.Builds.ByNames {
-			build := XrayWatchProjectResourcesElement{
-				Type:     "build",
-				Name:     byName.Name,
-				BinMgrID: byName.BinMgrID,
-			}
-
-			payloadBody.ProjectResources.Resources = append(payloadBody.ProjectResources.Resources, build)
-		}
-	}
-
-	return nil
-}
-
-func ConfigureBundles(payloadBody *XrayWatchBody, params XrayWatchParams) error {
-	// to be implemented
-	return nil
-}
-
-func (xws *XrayWatchService) Update(params XrayWatchParams) error {
-	payloadBody, err := CreateBody(params)
+func (xws *XrayWatchService) Update(params utils.XrayWatchParams) error {
+	payloadBody, err := utils.CreateBody(params)
 
 	// the update payload must not have a name
 	payloadBody.GeneralData.Name = ""
@@ -270,7 +110,7 @@ func (xws *XrayWatchService) Update(params XrayWatchParams) error {
 	}
 
 	httpClientsDetails := xws.XrayDetails.CreateHttpClientDetails()
-	utils.SetContentType("application/json", &httpClientsDetails.Headers)
+	artUtils.SetContentType("application/json", &httpClientsDetails.Headers)
 	var url = xws.GetXrayWatchUrl() + "/" + params.Name
 	var resp *http.Response
 	var respBody []byte
@@ -293,47 +133,44 @@ func (xws *XrayWatchService) Update(params XrayWatchParams) error {
 	return nil
 }
 
-func (xws *XrayWatchService) Get(watchName string) (watchResp *XrayWatchParams, err error) {
+func (xws *XrayWatchService) Get(watchName string) (watchResp *utils.XrayWatchParams, err error) {
 	httpClientsDetails := xws.XrayDetails.CreateHttpClientDetails()
 	log.Info("Getting watch...")
 	resp, body, _, err := xws.client.SendGet(xws.GetXrayWatchUrl()+"/"+watchName, true, &httpClientsDetails)
-	watch := XrayWatchBody{}
+	watch := utils.XrayWatchBody{}
 
 	if err != nil {
-		return &XrayWatchParams{}, err
+		return &utils.XrayWatchParams{}, err
 	}
 	if resp.StatusCode != http.StatusOK {
-		return &XrayWatchParams{}, errorutils.CheckError(errors.New("Artifactory response: " + resp.Status + "\n" + clientutils.IndentJson(body)))
+		return &utils.XrayWatchParams{}, errorutils.CheckError(errors.New("Artifactory response: " + resp.Status + "\n" + clientutils.IndentJson(body)))
 	}
 
 	err = json.Unmarshal(body, &watch)
 
 	if err != nil {
-		return &XrayWatchParams{}, errors.New("failed unmarshalling watch " + watchName)
+		return &utils.XrayWatchParams{}, errors.New("failed unmarshalling watch " + watchName)
 	}
 
-	result := XrayWatchParams{
-		Name:        watch.GeneralData.Name,
-		Description: watch.GeneralData.Description,
-		Active:      watch.GeneralData.Active,
-		Repositories: XrayWatchRepositoriesParams{
-			Type:         "",             // WatchRepositoriesType
-			All:          XrayWatchAll{}, // XrayWatchAll
-			Repositories: map[string]XrayWatchRepository{},
-			XrayWatchPathFilters: XrayWatchPathFilters{
-				ExcludePatterns: []string{},
-				IncludePatterns: []string{},
-			},
+	result := NewXrayWatchParams()
+	result.Name = watch.GeneralData.Name
+	result.Description = watch.GeneralData.Description
+	result.Active = watch.GeneralData.Active
+	result.Repositories = utils.XrayWatchRepositoriesParams{
+		All:          utils.XrayWatchAll{},
+		Repositories: map[string]utils.XrayWatchRepository{},
+		XrayWatchPathFilters: utils.XrayWatchPathFilters{
+			ExcludePatterns: []string{},
+			IncludePatterns: []string{},
 		},
-		Builds: XrayWatchBuildsParams{
-			Type:    "", //       WatchBuildType
-			All:     XrayWatchBuildsAllParams{},
-			ByNames: map[string]XrayWatchBuildsByNameParams{},
-		},
-		Policies: watch.AssignedPolicies,
 	}
+	result.Builds = utils.XrayWatchBuildsParams{
+		All:     utils.XrayWatchBuildsAllParams{},
+		ByNames: map[string]utils.XrayWatchBuildsByNameParams{},
+	}
+	result.Policies = watch.AssignedPolicies
 
-	unpackWatchBody(&result, &watch)
+	utils.UnpackWatchBody(&result, &watch)
 
 	log.Debug("Artifactory response:", resp.Status)
 	log.Info("Done getting watch.")
@@ -341,215 +178,17 @@ func (xws *XrayWatchService) Get(watchName string) (watchResp *XrayWatchParams, 
 	return &result, nil
 }
 
-func unpackWatchBody(watch *XrayWatchParams, body *XrayWatchBody) {
-	for _, resource := range body.ProjectResources.Resources {
-		if resource.Type == "all-repos" {
-			watch.Repositories.Type = WatchRepositoriesAll
-			unpackFilters(resource.StringFilters, &watch.Repositories.All.Filters, &watch.Repositories)
-		}
-		if resource.Type == "repository" {
-			watch.Repositories.Type = WatchRepositoriesByName
-			repository := XrayWatchRepository{
-				Name:     resource.Name,
-				BinMgrID: resource.BinMgrID,
-			}
-			unpackFilters(resource.StringFilters, &repository.Filters, &watch.Repositories)
-			watch.Repositories.Repositories[repository.Name] = repository
-		}
-		if resource.Type == "all-builds" {
-			watch.Builds.Type = WatchBuildAll
-			watch.Builds.All.BinMgrID = resource.BinMgrID
-
-			for _, filter := range resource.StringFilters {
-				if filter.Type == "ant-patterns" {
-					pathFilters := filter.Value.(map[string]interface{})
-
-					if pathFilters["ExcludePatterns"] != nil {
-						for _, path := range pathFilters["ExcludePatterns"].([]interface{}) {
-							watch.Builds.All.ExcludePatterns = append(watch.Builds.All.ExcludePatterns, path.(string))
-						}
-					}
-					if pathFilters["IncludePatterns"] != nil {
-						for _, path := range pathFilters["IncludePatterns"].([]interface{}) {
-							watch.Builds.All.IncludePatterns = append(watch.Builds.All.IncludePatterns, path.(string))
-						}
-					}
-				}
-			}
-
-		}
-		if resource.Type == "build" {
-			watch.Builds.Type = WatchBuildByName
-			watch.Builds.ByNames[resource.Name] = XrayWatchBuildsByNameParams{
-				Name:     resource.Name,
-				BinMgrID: resource.BinMgrID,
-			}
-		}
-	}
-
-	// Sort all the properties so they are returned in a consistent format
-
-	sort.Strings(watch.Repositories.ExcludePatterns)
-	sort.Strings(watch.Repositories.IncludePatterns)
+func NewXrayWatchParams() utils.XrayWatchParams {
+	return utils.XrayWatchParams{}
 }
 
-func unpackFilters(filters []XrayWatchFilter, output *XrayWatchFilters, repos *XrayWatchRepositoriesParams) {
-
-	for _, filter := range filters {
-		if filter.Type == "package-type" {
-			output.PackageTypes = append(output.PackageTypes, filter.Value.(string))
-		}
-		if filter.Type == "regex" {
-			output.Names = append(output.Names, filter.Value.(string))
-		}
-		if filter.Type == "path-regex" {
-			output.Paths = append(output.Paths, filter.Value.(string))
-		}
-		if filter.Type == "mime-type" {
-			output.MimeTypes = append(output.MimeTypes, filter.Value.(string))
-		}
-		if filter.Type == "property" {
-			output.Properties = map[string]string{}
-			filterParams := filter.Value.(map[string]interface{})
-			key := filterParams["key"].(string)
-			value := filterParams["value"].(string)
-			output.Properties[key] = value
-		}
-
-		if filter.Type == "path-ant-patterns" {
-			// The path filters are defined once for repositories, either all, or by name
-			// So, we only add the paths once
-
-			pathFilters := filter.Value.(map[string]interface{})
-
-			if len(repos.ExcludePatterns) == 0 && pathFilters["ExcludePatterns"] != nil {
-				for _, path := range pathFilters["ExcludePatterns"].([]interface{}) {
-					repos.ExcludePatterns = append(repos.ExcludePatterns, path.(string))
-				}
-			}
-			if len(repos.IncludePatterns) == 0 && pathFilters["IncludePatterns"] != nil {
-				for _, path := range pathFilters["IncludePatterns"].([]interface{}) {
-					repos.IncludePatterns = append(repos.IncludePatterns, path.(string))
-				}
-			}
-		}
-	}
-
-	// Sorting so that outputs are consistent
-	// Not sure if this is the best solution.
-	sort.Strings(output.PackageTypes)
-	sort.Strings(output.Names)
-	sort.Strings(output.Paths)
-	sort.Strings(output.MimeTypes)
+func NewXrayPolicy() utils.XrayPolicy {
+	return utils.XrayPolicy{}
 }
 
-func NewXrayWatchParams() XrayWatchParams {
-	return XrayWatchParams{}
-}
-
-type XrayWatchParams struct {
-	Name        string
-	Description string
-	Active      bool
-
-	Repositories XrayWatchRepositoriesParams
-
-	Builds   XrayWatchBuildsParams
-	Policies []XrayWatchPolicy
-}
-
-type XrayWatchBody struct {
-	GeneralData      XrayWatchGeneralParams    `json:"general_data"`
-	ProjectResources XrayWatchProjectResources `json:"project_resources,omitempty"`
-	AssignedPolicies []XrayWatchPolicy         `json:"assigned_policies,omitempty"`
-}
-
-type WatchBuildType string
-type WatchRepositoriesType string
-
-type XrayWatchGeneralParams struct {
-	ID          string `json:"id,omitempty"`
-	Name        string `json:"name,omitempty"` // Must be empty on update.
-	Description string `json:"description"`
-	Active      bool   `json:"active"`
-}
-
-type XrayWatchRepositoriesParams struct {
-	Type         WatchRepositoriesType
-	All          XrayWatchAll
-	Repositories map[string]XrayWatchRepository
-	XrayWatchPathFilters
-}
-
-type XrayWatchAll struct {
-	Filters XrayWatchFilters
-}
-
-type XrayWatchFilters struct {
-	PackageTypes []string
-	Names        []string
-	Paths        []string
-	MimeTypes    []string
-	Properties   map[string]string
-}
-
-type XrayWatchBuildsParams struct {
-	Type    WatchBuildType
-	All     XrayWatchBuildsAllParams
-	ByNames map[string]XrayWatchBuildsByNameParams
-}
-
-type XrayWatchBuildsAllParams struct {
-	BinMgrID string `json:"bin_mgr_id"`
-	XrayWatchPathFilters
-}
-
-type XrayWatchBuildsByNameParams struct {
-	Name     string
-	BinMgrID string
-}
-
-type XrayWatchFilter struct {
-	Type  string      `json:"type"`
-	Value interface{} `json:"value"`
-}
-
-type XrayWatchFilterPropertyValue struct {
-	Key   string `json:"key"`
-	Value string `json:"value"`
-}
-
-type XrayWatchProjectResources struct {
-	Resources []XrayWatchProjectResourcesElement `json:"resources"`
-}
-
-type XrayWatchProjectResourcesElement struct {
-	Name          string            `json:"name,omitempty"`
-	BinMgrID      string            `json:"bin_mgr_id,oitempty"`
-	Type          string            `json:"type"`
-	StringFilters []XrayWatchFilter `json:"filters,omitempty"`
-}
-
-type XrayWatchRepository struct {
-	Name          string            `json:"name"`
-	StringFilters []XrayWatchFilter `json:"filters"`
-	BinMgrID      string            `json:"bin_mgr_id"`
-	Filters       XrayWatchFilters
-}
-
-type XrayWatchPathFilters struct {
-	ExcludePatterns []string `json:"ExcludePatterns"`
-	IncludePatterns []string `json:"IncludePatterns"`
-}
-
-func NewXrayWatchRepository(name string, binMgrID string) XrayWatchRepository {
-	return XrayWatchRepository{
+func NewXrayWatchRepository(name string, binMgrID string) utils.XrayWatchRepository {
+	return utils.XrayWatchRepository{
 		Name:     name,
 		BinMgrID: binMgrID,
 	}
-}
-
-type XrayWatchPolicy struct {
-	Name string `json:"name"`
-	Type string `json:"type"`
 }
